# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "description": "Infrastructure inventory, observation, incident response, execution, and access control API.",
    "title": "KloudView API",
    "version": "0.1.0"
  },
  "servers": [
    {
      "description": "Same-origin Server",
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "tags": [
          "users"
        ],
        "operationId": "getApiV1Users",
        "summary": "GET /api/v1/users",
        "responses": {
          "2XX": {
            "description": "Successful response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error response"
          }
        },
        "security": [
          {
            "headerSubject": []
          }
        ]
      },
      "post": {
        "tags": [
          "users"
        ],
        "operationId": "postApiV1Users",
        "summary": "POST /api/v1/users",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "2XX": {
            "description": "Successful response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error response"
          }
        },
        "security": [
          {
            "headerSubject": []
          }
        ]
      }
    },
    "/api/v1/users/{id}": {
      "delete": {
        "tags": [
          "users"
        ],
        "operationId": "deleteApiV1UsersId",
        "summary": "DELETE /api/v1/users/{id}",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "2XX": {
            "description": "Successful response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error response"
          }
        },
        "security": [
          {
            "headerSubject": []
          }
        ]
      },
      "put": {
        "tags": [
          "users"
        ],
        "operationId": "putApiV1UsersId",
        "summary": "PUT /api/v1/users/{id}",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "2XX": {
            "description": "Successful response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error response"
          }
        },
        "security": [
          {
            "headerSubject": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "properties": {
          "error": {
            "type": "object"
          }
        },
        "type": "object"
      }
    },
    "securitySchemes": {
      "agentBearer": {
        "scheme": "bearer",
        "type": "http"
      },
      "headerSubject": {
        "in": "header",
        "name": "X-KloudView-Subject",
        "type": "apiKey"
      }
    }
  }
}